refactor: useRouer in page folder
